+++ packages/api-client/src/team/feature/FeatureAPI.ts
@@ -33,6 +33,8 @@ import {
   FeatureMLSE2EId,
   FeatureMLSMigration,
   FeatureDownloadPath,
+  FeatureDomainRegistration,
+  FeatureChannels,
 } from './Feature';
 import {InvalidAppLockTimeoutError} from './FeatureError';
 import {FeatureList} from './FeatureList';
@@ -50,7 +52,9 @@ export class FeatureAPI {
     CALLING_VIDEO: 'videoCalling',
     SELF_DELETING_MESSAGES: 'selfDeletingMessages',
     DIGITAL_SIGNATURES: 'digitalSignatures',
+    DOMAIN_REGISTRATION: 'domainRegistration',
     DL_PATH: 'enforceFileDownloadLocation',
+    CHANNELS: 'channels',
     CONVERSATION_GUEST_LINKS: 'conversationGuestLinks',
     FEATURE_CONFIGS: '/feature-configs',
     FEATURES: 'features',
@@ -347,6 +351,17 @@ export class FeatureAPI {
     return response.data;
   }
 
+  // Get domainRegistration feature
+  public async getDomainRegistrationFeature(teamId: string): Promise<FeatureDomainRegistration> {
+    const config: AxiosRequestConfig = {
+      method: 'get',
+      url: `${FeatureAPI.URL.TEAMS}/${teamId}/${FeatureAPI.URL.FEATURES}/${FeatureAPI.URL.DOMAIN_REGISTRATION}`,
+    };
+
+    const response = await this.client.sendJSON<FeatureDomainRegistration>(config);
+    return response.data;
+  }
+
   public async getAppLockFeature(teamId: string): Promise<FeatureAppLock> {
     const config: AxiosRequestConfig = {
       method: 'get',
@@ -393,6 +408,17 @@ export class FeatureAPI {
     };
 
     const response = await this.client.sendJSON<FeatureDownloadPath>(config);
+    return response.data;
+  }
+
+  public async putChannelsFeature(teamId: string, channelFeature: Partial<FeatureChannels>): Promise<FeatureChannels> {
+    const config: AxiosRequestConfig = {
+      data: channelFeature,
+      method: 'put',
+      url: `${FeatureAPI.URL.TEAMS}/${teamId}/${FeatureAPI.URL.FEATURES}/${FeatureAPI.URL.CHANNELS}`,
+    };
+
+    const response = await this.client.sendJSON<FeatureChannels>(config);
     return response.data;
   }
 }